chore: added disclaimer
